test: rename from jest to test
