+++ packages/@lwc/module-resolver/scripts/test/matchers/to-throw-error-with-type.ts
