+++ packages/@lwc/module-resolver/src/__tests__/multi-version.spec.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import '../../scripts/jest/types';
+import '../../scripts/test/types';
 import { resolveModule } from '../index';
 import { RegistryType } from '../types';
 import { fixture, NO_LWC_MODULE_FOUND_CODE } from './test-utils';
